+++ src/test/kotlin/com/pestphp/pest/inspections/InvalidTestNameCaseInspectionTest.kt
@@ -1,7 +1,9 @@
 package com.pestphp.pest.inspections
 
+import com.intellij.testFramework.TestDataPath
 import com.pestphp.pest.PestLightCodeFixture
 
+@TestDataPath("\$CONTENT_ROOT/resources/com/pestphp/pest/inspections")
 class InvalidTestNameCaseInspectionTest : PestLightCodeFixture() {
     override fun getTestDataPath(): String {
         return "src/test/resources/com/pestphp/pest/inspections"
@@ -32,5 +34,17 @@ class InvalidTestNameCaseInspectionTest : PestLightCodeFixture() {
         myFixture.getAllQuickFixes().forEach { myFixture.launchAction(it) }
 
         myFixture.checkResultByFile("InvalidTestNameCase.after.php")
+    }
+
+    fun testValidTestNameWhenWrongCasingOnSomeWords() {
+        myFixture.configureByFile("ValidTestNameWhenWrongCasingOnOneWord.php")
+
+        myFixture.checkHighlighting()
+    }
+
+    fun testValidTestNameCaseInHigherOrderTest() {
+        myFixture.configureByFile("ValidHigherOrderTestNameCase.php")
+
+        myFixture.checkHighlighting()
     }
 }