+++ src/test/resources/com/pestphp/pest/datasets/DatasetTest.php
@@ -1,0 +1,12 @@
+<?php
+
+dataset('dojos', [
+    'Cobra Kai',
+    'Eagle Fang',
+    'Miyagi Do'
+]);
+
+
+it('can check if in the valley', function (string $dojo) {
+
+})->with('<caret>');