+++ src/test/resources/com/pestphp/pest/datasets/Datasets.php
@@ -1,0 +1,18 @@
+<?php
+
+dataset('stances', [
+    'Seiza',
+    'Musubi Dachi',
+    'Heisoku Dachi',
+    'Heiko Dachi',
+    'Zenkutsu Dachi',
+    'Han Zenkutsu Dachi',
+    'Kokutsu Dachi',
+]);
+
+
+
+
+
+
+