+++ src/test/resources/com/pestphp/pest/inspections/ValidHigherOrderTestNameCase.php
@@ -1,0 +1,5 @@
+<?php
+
+test('basic test')
+    ->expect(true)
+    ->toBeTrue();