+++ src/test/resources/com/pestphp/pest/inspections/ValidTestNameWhenWrongCasingOnOneWord.php
@@ -1,0 +1,13 @@
+<?php
+
+test('can be used with camelCase', function () {
+    $this->assertTrue(true);
+});
+
+test('can be used with snake_case or UPPER_CASE', function () {
+    $this->assertTrue(true);
+});
+
+test('can "quote_out" the word', function () {
+    $this->assertTrue(true);
+});